Replace OrderContentsController#create with OrdersController#populate
Based on https://github.com/solidusio/solidus/pull/4102.
