+++ templates/app/controllers/spree/orders_controller.rb
@@ -46,38 +46,6 @@ module Spree
       end
     end
 
-    # Adds a new item to the order (creating a new order if none already exists)
-    def populate
-      @order = current_order(create_order_if_necessary: true)
-      authorize! :update, @order, cookies.signed[:guest_token]
-
-      variant  = Spree::Variant.find(params[:variant_id])
-      quantity = params[:quantity].present? ? params[:quantity].to_i : 1
-
-      # 2,147,483,647 is crazy. See issue https://github.com/spree/spree/issues/2695.
-      if !quantity.between?(1, 2_147_483_647)
-        @order.errors.add(:base, t('spree.please_enter_reasonable_quantity'))
-      else
-        begin
-          @line_item = @order.contents.add(variant, quantity)
-        rescue ActiveRecord::RecordInvalid => error
-          @order.errors.add(:base, error.record.errors.full_messages.join(", "))
-        end
-      end
-
-      respond_with(@order) do |format|
-        format.html do
-          if @order.errors.any?
-            flash[:error] = @order.errors.full_messages.join(", ")
-            redirect_back_or_default(spree.root_path)
-            return
-          else
-            redirect_to cart_path
-          end
-        end
-      end
-    end
-
     def empty
       if @order = current_order
         authorize! :update, @order, cookies.signed[:guest_token]
