+++ templates/config/routes.rb
@@ -36,6 +36,8 @@ Spree::Core::Engine.routes.draw do
 
   resources :products, only: [:index, :show]
 
+  resources :order_contents, only: :create
+
   get '/locale/set', to: 'locale#set'
   post '/locale/set', to: 'locale#set', as: :select_locale
 
@@ -49,7 +51,6 @@ Spree::Core::Engine.routes.draw do
   get '/orders/:id/token/:token' => 'orders#show', as: :token_order
 
   resources :orders, except: [:index, :new, :create, :destroy] do
-    post :populate, on: :collection
     resources :coupon_codes, only: :create
   end
