+++ templates/spec/system/coupon_code_spec.rb
@@ -89,6 +89,7 @@ RSpec.describe 'Coupon code promotions', type: :system, js: true do
           allow_any_instance_of(Spree::CheckoutController).to receive_messages(spree_current_user: user)
           allow_any_instance_of(Spree::OrdersController).to receive_messages(spree_current_user: user)
           allow_any_instance_of(Spree::CouponCodesController).to receive_messages(spree_current_user: user)
+          allow_any_instance_of(Spree::OrderContentsController).to receive_messages(spree_current_user: user)
         end
 
         context 'with saved credit card' do
